fix: regression for cte reference
